refactor(state): extract RecentPreflightStore and MutationAuditLog from state.rs
Reduce state.rs from 1458 to 1378 lines by extracting two isolated
concerns into dedicated modules: preflight_store.rs (99 lines) for
recent preflight management and mutation_audit.rs (76 lines) for
audit logging. AppState delegates via thin wrapper methods, preserving
the existing public API with zero test regressions.
