+++ crates/codelens-mcp/src/main.rs
@@ -5,7 +5,9 @@ mod dispatch;
 mod dispatch_access;
 mod dispatch_response;
 mod error;
+mod mutation_audit;
 mod mutation_gate;
+mod preflight_store;
 mod prompts;
 mod protocol;
 mod resource_analysis;
